requirements/dev.txt
-r common.txt

# Django Modules
django-debug-toolbar
django-extensions

# Other Modules
markdown
setuptools
twine
wheel